Fix: disable @next/next/no-html-link-for-pages
This commit disables the ESLint rule @next/next/no-html-link-for-pages
on two specific lines. We're disabling the rule because we need the <a>
tags there and the error caused by this rule is preventing the
application from building.

The reason we need the <a> tags because the NextJS <Link>
component is only for client-side page transitions, which we are not doing
here. These links specifically point to the backend for redirects to
Auth0-hosted pages, and thus need to be regular HTML links.